Remove useThing composable
